feat : exception / NotFoundProvider
api 로 provider(ex : github, google ... ) 에 대해 로그인 요청이 왔을 경우 해당 provider 가 존재하지 않는다면 에러가 발생하도록 하고, 에러가 정상적으로 발생하는지 테스트 하였습니다.
