+++ moa3/src/main/java/com/moa/moa3/exception/oauth/NotFoundProviderException.java
@@ -1,0 +1,11 @@
+package com.moa.moa3.exception.oauth;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class NotFoundProviderException extends RuntimeException{
+    public NotFoundProviderException(String providerName) {
+        super("NotFoundProvider: " + providerName);
+        log.error("NotFoundProvider: " + providerName);
+    }
+}
