+++ moa3/src/main/java/com/moa/moa3/service/oauth/OAuthService.java
@@ -1,12 +1,19 @@
 package com.moa.moa3.service.oauth;
 
+import com.moa.moa3.dto.oauth.OAuthProvider;
 import com.moa.moa3.dto.oauth.UserData;
+import com.moa.moa3.util.oauth.OAuthProviderFactory;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 public class OAuthService {
 
+    private final OAuthProviderFactory oAuthProviderFactory;
+
     public UserData getUserData(String provider, String code) {
+        OAuthProvider oAuthProvider = oAuthProviderFactory.getProvider(provider);
         return new UserData("lorem", "ipsum", "dolor");
     }
 }
